pyproject.toml
[project]
name = "to-vector"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pika>=1.3.2",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]
